Time: 0 ms (100.00%) | Memory: 27.3 MB (15.50%) - LeetSync

+++ 344-reverse-string/reverse-string.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     void reverseString(vector<char>& s) {
-        long int left = 0;
-        long int right = s.size() - 1;
+        int left = 0;
+        int right = s.size() - 1;
 
         while(left < right){
             swap(s[left], s[right]);
